test(theme-light): extract WCAG contrast helpers to shared test-utils
